issue #32 move files
